New option --normalize-dist to normalize tree distance into [0,1]

+++ tree/mtreeset.h
@@ -162,7 +162,7 @@ public:
 		@param mode RF_ALL_PAIR or RF_ADJACENT_PAIR
 		@param weight_threshold minimum weight cutoff
 	*/
-	void computeRFDist(int *rfdist, int mode = RF_ALL_PAIR, double weight_threshold = -1000);
+	void computeRFDist(double *rfdist, int mode = RF_ALL_PAIR, double weight_threshold = -1000);
 
 	/**
 		compute the Robinson-Foulds distance between trees
@@ -171,8 +171,8 @@ public:
         @param k_by_k true to compute distances between corresponding k-th tree of two tree sets,
             false to do all-by-all
 	*/
-	void computeRFDist(int *rfdist, MTreeSet *treeset2, bool k_by_k,
-		const char* info_file = NULL, const char *tree_file = NULL, int *incomp_splits = NULL);
+	void computeRFDist(double *rfdist, MTreeSet *treeset2, bool k_by_k,
+		const char* info_file = NULL, const char *tree_file = NULL, double *incomp_splits = NULL);
 
 	int categorizeDistinctTrees(IntVector &category);
